FIX: Thread from MCMC was not returning correctly DEL: Removed the -t command line option and it's implementation ADD: registerEstimable(string, vector<double>) to estimate a whole double. ADD: report prefixes for minimisation and simulations FIX: FMM was printing out a long double without converting it. This is broken in compiler. CHG: Layer report now uses PARAM_YEARS instead of PARAM_YEAR DEL: Removed all of the threading code in the MCMC system CHG: MCMC now doesn't try to call choleskyDecomp or build covariance every iteration CHG: MCMC is now mostly implemented, but is not yet functional

+++ src/Minimizers/GammaDiff/Engine/FMM.cpp
@@ -308,7 +308,7 @@ void FMM::fMin(vector<double>& Candidates, double& Score, vector<double>& Gradie
     // Is StepSize too Small?
     if (dLambda < dLambdaMin) {
       iRet = -3;
-      cerr << FMM_SMALL_LINEAR_STEP_SIZE << " (" << dLambda << ")" << endl;
+      cerr << FMM_SMALL_LINEAR_STEP_SIZE << " (" << (double)dLambda << ")" << endl;
       return;
     }
